Add parameter to change video speed
It seems like the two cameras were shot on ever so slightly different
frame rates (or maybe there is a different problem) because the videos
from the two cameras seems to drift as time goes on. This commit makes
it possible to change the video speed by some factor (I've figured out
what factor to apply to the maincam so that it stays in sync with the
sidecam - hopefully it works!).